refactored to follow the Single responsibility principle, and added some classes
