+++ CmsLibrary/DataAccess/CostMonitoring/ProjectSelection/MainProjectsSqlConnector.cs
@@ -16,50 +16,51 @@ namespace CmsLibrary.DataAccess.CostMonitoring {
         public void CreateMainProject( MainProjectsModel credentials , string events , string tableName ) {
             using( SqlConnection connection = new SqlConnection( GlobalConfig.ConnString ) )
             {
-                using( SqlCommand cmd = new SqlCommand("dbo.spProjects",connection) )
+                using( SqlCommand cmd = new SqlCommand( "dbo.spProjects" , connection) )
                 {
                     cmd.CommandType = CommandType.StoredProcedure;
-                   
-                    
-                    SqlParameter Id = new SqlParameter( "@Id" , SqlDbType.Int );
-                    Id.Direction = ParameterDirection.Output;
-
                     cmd.Parameters.AddWithValue( "@event" , events );
                     cmd.Parameters.AddWithValue( "@tableName" , tableName );
                     cmd.Parameters.AddWithValue( "@ProjectName" , credentials.ProjectName );
                     cmd.Parameters.AddWithValue( "@DateStarted" , credentials.DateStarted );
-                    cmd.Parameters.Add(  Id );
+
+                    //SqlParameter p = new SqlParameter( );
+                    //p.ParameterName = "@Id";
+                    //p.Size = 4;
+                    //p.Direction = ParameterDirection.Output;
+                    //cmd.Parameters.Add( p );
+
                     cmd.Connection.Open( );
                     cmd.ExecuteNonQuery( );
 
-                    if( Id.Value !=DBNull.Value )
-                    {
-                        int pkId = Convert.ToInt32( Id.Value );
-                        Console.WriteLine( pkId.ToString( ) );
-                    }
-                    else
-                    {
-                        System.Windows.Forms.MessageBox.Show( "Test" );
-                    }
+                    //int id = Convert.ToInt32( cmd.ExecuteScalar( ) );
+                    //Console.WriteLine( id );
+
                 }
             }
         }
 
-        public void RemoveProject( IProjectsCredentials Id ) {
-            throw new NotImplementedException( );
-        }
-
-        public DataTable tableProjects( ) {
-            throw new NotImplementedException( );
-        }
-
-        public DataTable tableSubProjects( ) {
+        public DataTable tableMainProjects( ) {
             throw new NotImplementedException( );
         }
 
         public void UpdateMainProject( MainProjectsModel credentials ) {
             throw new NotImplementedException( );
         }
+
+        //public void RemoveProject( IProjectsCredentials Id ) {
+        //    throw new NotImplementedException( );
+        //}
+
+        //public DataTable tableProjects( ) {
+        //    throw new NotImplementedException( );
+        //}
+
+        //public DataTable tableSubProjects( ) {
+        //    throw new NotImplementedException( );
+        //}
+
+
 
     }
 }
