+++ CmsLibrary/DataAccess/CostMonitoring/ProjectSelection/SubProjectsSqlConnector.cs
@@ -12,7 +12,8 @@ using System.Data.SqlClient;
 
 namespace CmsLibrary.DataAccess.CostMonitoring {
     public class SubProjectsSqlConnector : ISubProjectsConnection {
-        public void CreateMainProject( SubProjectsModel credentials , string events , string tableName ) {
+
+        public void CreateSubProject( SubProjectsModel credentials , string events , string tableName ) {
             using( SqlConnection connection = new SqlConnection(GlobalConfig.ConnString))
             {
                 using( SqlCommand cmd = new SqlCommand("spProjects",connection) )
@@ -20,21 +21,17 @@ namespace CmsLibrary.DataAccess.CostMonitoring {
                     cmd.CommandType = CommandType.StoredProcedure;
                     cmd.Parameters.AddWithValue( "@event" , events );
                     cmd.Parameters.AddWithValue( "@tableName" , tableName );
-                    cmd.Parameters.AddWithValue( "@ProjectName" , credentials.ProjectName );
+                    cmd.Parameters.AddWithValue( "@SubProjectName" , credentials.ProjectName );
                     cmd.Parameters.AddWithValue( "@DateStarted" , credentials.DateStarted );
+                   
 
                     cmd.Connection.Open( );
+                    //int mainProjectId = Convert.ToInt32( cmd.ExecuteScalar( ) );
+                    //cmd.Parameters.AddWithValue( "@MainProjectId" , mainProjectId );
                     cmd.ExecuteNonQuery( );
+                    //Console.WriteLine( mainProjectId );
                 }
             }
-        }
-
-        public void RemoveProject( IProjectsCredentials Id ) {
-            throw new NotImplementedException( );
-        }
-
-        public DataTable tableProjects( ) {
-            throw new NotImplementedException( );
         }
 
         public DataTable tableSubProjects( ) {
@@ -44,5 +41,7 @@ namespace CmsLibrary.DataAccess.CostMonitoring {
         public void Update( SubProjectsModel credentials ) {
             throw new NotImplementedException( );
         }
+
+
     }
 }
