--- CmsUI/RevisionedUI/Cost monitoring/Bill Mat/bill_mat_UserControl.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace GSG_Builders.RevisionedUI.Bill_Mat {
-    public partial class bill_mat_UserControl : UserControl {
-        public bill_mat_UserControl ( ) {
-            InitializeComponent( );
-        }
-    }
-}
